+++ src/F1_plotting.py
@@ -4,6 +4,9 @@ import os
 import matplotlib.pyplot as plt
 import pandas as pd
 
+import networkx as nx
+import oemof.network.graph as graph
+
 from src.constants import (
     PATHS_TO_PLOTS,
     PLOTS_BUSSES,
@@ -12,15 +15,24 @@ from src.constants import (
     PLOTS_COSTS,
     PROJECT_DATA,
     LABEL,
+    OUTPUT_FOLDER,
     PATH_OUTPUT_FOLDER,
+    PLOTS_DEMANDS,
+    PLOTS_RESOURCES,
 )
+
 from src.constants_json_strings import (
     SIMULATION_SETTINGS,
     PROJECT_NAME,
     SCENARIO_NAME,
     KPI,
     KPI_COST_MATRIX,
+    ENERGY_CONSUMPTION,
+    TIMESERIES,
+    ENERGY_PRODUCTION,
 )
+
+from src.E1_process_results import convert_demand_to_dataframe
 
 r"""
 Module F1 describes all the functions that create plots.
@@ -39,20 +51,24 @@ def flows(dict_values, user_input, project_data, results_timeseries, bus, interv
     """
     Parameters
     ----------
-    user_input: dict
+    dict_values: dict,
+
+    user_input: dict,
         part of the dict_values that includes the output folders name
-    project_data: dict
+    project_data: dict,
         part of the dict_values that includes Name for setting title of plots
     results_timeseries: pd Dataframe
         Timeseries that is to be plotted
-    bus: str
-        sector that is to be plotted, ie. energyVectors of the energy system - not each and every bus.
-    interval: int
+    bus: str,
+        sector that is to be plotted, ie. energyVectors of the energy system -
+        - not each and every bus.
+    interval: int,
         Time interval in days covered on the x-axis
 
     Returns
     -------
-    Plot of "interval" duration on x-axis for all energy flows connected to the main bus supplying the demand of a specific sector
+    Plot of "interval" duration on x-axis for all energy flows connected to the main bus supplying
+    the demand of a specific sector
     """
 
     logging.info("Creating plots for %s sector, %s days", bus, interval)
@@ -113,10 +129,13 @@ def flows(dict_values, user_input, project_data, results_timeseries, bus, interv
 
 
 def capacities(dict_values, user_input, project_data, assets, capacities):
-    """Determines the assets for which the optimized capacity is larger than zero and then plots those capacities in a bar chart.
+    """Determines the assets for which the optimized capacity is larger than zero and then plots
+    those capacities in a bar chart.
 
     Parameters
     ----------
+    dict_values: dict
+        Dict with all simulation parameters
     user_input : dict
         Simulation settings including the output path
     project_data : dict
@@ -182,8 +201,9 @@ def capacities(dict_values, user_input, project_data, assets, capacities):
 
 def evaluate_cost_parameter(dict_values, parameter, file_name):
     """
-    Generates pie plot of a chosen cost parameter, and if one asset is overly present in the cost distribution with 90% of the costs,
-    a pie plot of the distribution of the remaining 10% of the costs.
+    Generates pie plot of a chosen cost parameter, and if one asset is overly present in the cost
+    distribution with 90% of the costs, a pie plot of the distribution of the remaining 10% of
+    the costs.
 
     Parameters
     ----------
@@ -284,11 +304,14 @@ def determine_if_plotting_necessary(parameter_values):
 
 def group_costs(costs, names):
     """
-    Calculates the percentage of different asset of the costs and also groups them by asset/DSO source/others
+    Calculates the percentage of different asset of the costs and also groups them by asset/DSO
+    source/others
     Parameters
     ----------
-    costs_perc: dict
-        dict relevant cost data, and asset name
+    costs: dict,
+        dict relevant cost data
+    names: dict,
+        dict with asset name
 
     Returns
     -------
@@ -324,7 +347,8 @@ def group_costs(costs, names):
 
 def recalculate_distribution_of_rest_costs(costs_perc_grouped_pandas):
     """
-    Determines whether there is one major player in the cost distribution, and if so, prepares plotting the remaining percent-
+    Determines whether there is one major player in the cost distribution, and if so, prepares
+    plotting the remaining percent-
 
     Parameters
     ----------
@@ -365,6 +389,8 @@ def plot_a_piechart(dict_values, settings, file_name, costs, label, title):
 
     Parameters
     ----------
+    dict_values: dict
+        Dict with all simulation parameters
     settings: dict
         includes output path
 
@@ -403,6 +429,115 @@ def plot_a_piechart(dict_values, settings, file_name, costs, label, title):
     return
 
 
+def convert_plot_data_to_dataframe(plot_data_dict, data_type):
+    """
+
+    Parameters
+    ----------
+    plot_data_dict: dict
+        timeseries for either demand or supply
+    data_type: str
+        one of "demand" or "supply"
+
+    Returns
+    -------
+    df: pandas:`pandas.DataFrame<frame>`,
+        timeseries for plotting
+    """
+    # Later, this dataframe can be passed to a function directly make the graphs with Plotly
+    df = pd.DataFrame.from_dict(plot_data_dict[data_type], orient="columns")
+
+    # Change the index of the dataframe
+    df.reset_index(level=0, inplace=True)
+    # Rename the timestamp column from 'index' to 'timestamp'
+    df = df.rename(columns={"index": "timestamp"})
+    return df
+
+
+def extract_plot_data_and_title(dict_values, df_dem=None):
+    """Dataframe used for the plots in the report
+
+    Parameters
+    ----------
+    dict_values: dict
+        output values of MVS
+
+    df_dem: :pandas:`pandas.DataFrame<frame>`
+        summarized demand information for each demand
+
+    Returns
+    -------
+    :pandas:`pandas.DataFrame<frame>`
+
+    """
+    if df_dem is None:
+        df_dem = convert_demand_to_dataframe(dict_values)
+
+    # Collect the keys of various resources (PV, Wind, etc.)
+    resources = dict_values[ENERGY_PRODUCTION]
+    res_keys = [k for k in resources.keys() if "DSO_" not in k]
+
+    # Gather all the keys of the various plots for later use in the graphOptions.csv
+    dict_for_plots = {"demands": {}, "supplies": {}}
+    dict_plot_labels = {}
+
+    # Add all the demands to the dict_for_plots dictionary, including the timeseries values
+    for demand in df_dem.Demands:
+        dict_for_plots["demands"].update(
+            {demand: dict_values[ENERGY_CONSUMPTION][demand][TIMESERIES]}
+        )
+        dict_plot_labels.update(
+            {demand: dict_values[ENERGY_CONSUMPTION][demand][LABEL]}
+        )
+
+    # Add all the resources to the dict_for_plots dictionary, including the timeseries values
+    for resource in res_keys:
+        dict_for_plots["supplies"].update(
+            {resource: dict_values[ENERGY_PRODUCTION][resource][TIMESERIES]}
+        )
+        dict_plot_labels.update(
+            {resource: dict_values[ENERGY_PRODUCTION][resource][LABEL]}
+        )
+
+    return dict_for_plots, dict_plot_labels
+
+
+def parse_simulation_log(path_log_file=None, log_type="ERROR"):
+    """Gather the log message of a certain type in a given log file
+
+    Parameters
+    ----------
+    path_log_file: str/None
+        path to the mvs log file
+        Default: None
+    log_type: str
+        one of "ERROR" or "WARNING"
+        Default: "ERROR"
+
+    Returns
+    -------
+
+    """
+    # Dictionaries to gather non-fatal warning and error messages that appear during the simulation
+    logs_dict = {}
+
+    if path_log_file is None:
+        path_log_file = os.path.join(OUTPUT_FOLDER, "mvs_logfile.log")
+
+    with open(path_log_file) as log_messages:
+        log_messages = log_messages.readlines()
+
+    i = 0
+    for line in log_messages:
+        if log_type in line:
+            i = i + 1
+            substrings = line.split(" - ")
+            message_string = substrings[-1]
+            logs_dict.update({i: message_string})
+
+    return logs_dict
+
+
 def draw_graph(
     dict_values,
     energysystem,
@@ -421,37 +556,33 @@ def draw_graph(
 
     Parameters
     ----------
-    energysystem :
-        param edge_labels:
-    node_color :
-        param edge_color: (Default value = "#eeac7e")
-    plot :
-        param node_size:
-    with_labels :
-        param arrows: (Default value = True)
-    layout :
-        return: (Default value = "dot")
-    user_input :
-        
-    edge_labels :
-         (Default value = True)
-    edge_color :
-         (Default value = "#eeac7e")
-    show_plot :
-         (Default value = True)
-    save_plot :
-         (Default value = True)
-    node_size :
-         (Default value = 5500)
-    arrows :
-         (Default value = True)
+    dict_values: dict
+        Dict with all simulation parameters
+    energysystem:
+    user_input: dict
+    edge_labels: bool
+        Default: True
+    node_color: str
+        Default: "#eeac7e"
+    edge_color: str
+        Default: "#eeac7e"
+    show_plot: bool
+        Default: True
+    save_plot: bool
+        Default: True
+    node_size: int
+        Default: 5500
+    with_labels: bool
+        Default: True
+    arrows: bool
+        Default: True
+    layout: str
+        Default: "dot"
 
     Returns
     -------
 
     """
-    import networkx as nx
-    import oemof.network.graph as graph
 
     # grph = graph.create_nx_graph(
     #    energysystem, filename=user_input[PATH_OUTPUT_FOLDER] + "network_graph.xml"
@@ -497,9 +628,6 @@ def draw_graph(
         )
 
 
-from src.constants import PLOTS_DEMANDS, PLOTS_RESOURCES
-
-
 def plot_input_timeseries(
     dict_values,
     user_input,
@@ -510,7 +638,8 @@ def plot_input_timeseries(
 ):
     r"""
     This function is called from C0 to plot the input timeseries provided to the MVS.
-    This includes demand profiles, generation profiles as well as timeseries for otherwise scalar values.
+    This includes demand profiles, generation profiles as well as timeseries for otherwise scalar
+    values.
 
     Parameters
     ----------