+++ CHANGELOG.md
@@ -42,6 +42,7 @@ Here is a template for new release sections
 - Deactivated function `C2.fuel_price_present_value` as it is not used and TBD (#495)
 - Modified the doc-strings in the definitions of some functions to abide by the formatting rules of numpy doc-strings (#505)
 - Suppressed the log messages of the Flask server (for report webapp) (#509) 
+- Move bulk data preparation code for report from F2 into E1 and F1 modules and into functions (#511)
 
 ### Removed
 -